pyproject.toml
[project]
name = "fastapiproject"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aio-pika==9.4.2",
    "aiomysql>=0.3.2",
    "apscheduler==3.11.0",
    "arrow>=1.4.0",
    "asgi-correlation-id>=4.3.4",
    "asyncpg>=0.31.0",
    "fastapi>=0.121.3",
    "fastapi-utils>=0.7.0,<1.0.0",
    "greenlet==3.2.3",
    "gunicorn==20.1.0",
    "loguru>=0.7.3",
    "motor==3.4.0",
    "prometheus-client==0.23.1",
    "pydantic-settings>=2.12.0",
    "pymongo==4.7.1",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.3",
    "redis==5.0.6",
    "requests>=2.32.5",
    "sqlmodel==0.0.14",
    "typing-inspect>=0.9.0",
    "uvicorn==0.34.3",
]
